Grenades and tigers and bears, oh my!
Implemented the first projectile weapon, the MGL 32. Fires grenades that bounce twice and explode, or explode on impact with something damageable.
Also got the burst fire pistol, the 93R in the game. Its pretty fun to shoot. Need to get some targets to shoot in soon.
I need to test it with people soon too.