Corrected behaviour, new spawn behaviour
Busy day today, clearly!
* Projectiles now have the expected behaviour. The ProtoGL's projectiles now disable their colliders when they explode, so you can't walk into them and make them explode 100 times.
* Repaired automatic/semi-automatic firing. Somewhere down the line I must've broken it, yarg
* Reworked weapon spawning - now uses a NetworkList<NetworkBehaviourReference> that is interpreted as BaseEquipment instances so that late-joiners can see players weapons correctly.
 * Weapons are cleaned up and new ones are spawned when you respawn via the respawn button. When revived, you shoud keep your equipment.
 * You now also spawn with two weapons - the Revolver and the ProtoGL.
* Added idle pose to Revolver and ProtoGL. This correctly changes on all clients when swapping, and is correctly set on late-joiners' ends too.
* Training Dummies now take damage from explosives.
* Debuffs should apply when correctly set on the weapon/projectile.

Notes:
Projectiles hold a reference to the weapon that fires them, or rather to that weapon's prefab. I will change the projectile so that stats are pulled from that, rather than having duplicate stats on the weapon. There won't ever be a case where a projectile does less radial damage when it travels further. This just doesn't make sense. And there won't be a case where a projectile that uses range-based damage will have radial damage too. It'll just hit what it collides with, and deal damage to that. There's an odd issue where hitting a rigidbody with a projectile can cause it to be launched, and I will look into this.